Blockchain-Node/requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.6
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.7
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
cssselect==1.1.0
decorator==5.0.7
dicttoxml==1.7.4
Django==2.2.6
django-cors-headers==3.1.1
docutils==0.17.1
elasticsearch==7.12.1
future==0.18.2
geocoder==1.38.1
geoip2==4.1.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.0.1
ip2geotools==0.1.5
IP2Location==8.5.1
isort==5.8.0
jeepney==0.6.0
keyring==23.0.1
lazy-object-proxy==1.6.0
lxml==4.6.3
maxminddb==2.0.3
mccabe==0.6.1
mfrc522==0.0.7
multidict==5.1.0
packaging==20.9
pip-review==1.1.0
pkginfo==1.7.0
publicip==1.0.1
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
pylint==2.8.2
pymongo==3.11.3
pyparsing==2.4.7
pyquery==1.4.3
pytz==2019.3
ratelim==0.1.6
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
RPi.GPIO==0.7.0
SecretStorage==3.3.1
selenium==3.141.0
six==1.15.0
spidev==3.5
sqlparse==0.3.0
toml==0.10.2
tqdm==4.60.0
twine==3.4.1
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==3.10.0.0
urllib3==1.26.4
webencodings==0.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
